requirements.txt
Flask
flask-sqlalchemy
Flask-WTF
Pillow